Removed multi transfer part from driver (#1)
Removed the multi transfer part from the driver. This part was unused and had some janky code (it was using a KEVENT as storage for how many bytes needed to be transfered). The USB Chief application only sends requests up to 64000 bytes. This is exactly the length the driver was checking for so it is not needed.

Tested the output of this and everything still works as expected (tested normal recordings and updating the firmware and busengine)